requirements.txt
streamlit
yfinance
fredapi
matplotlib
pandas
python-dotenv
